some optimization
